# Conflicts:
#	apps/cli/src/commands/mcp.rs
#	crates/cockpit-core/src/daemon/server/dispatch.rs
#	crates/cockpit-core/src/daemon/server/tests.rs
#	crates/cockpit-tui/src/tui/app/async_actions.rs
#	crates/cockpit-tui/src/tui/app/export_actions.rs
#	crates/cockpit-tui/src/tui/app/input.rs
#	crates/cockpit-tui/src/tui/app/local_commands.rs
#	crates/cockpit-tui/src/tui/settings/disk_daemon_fake.rs
#	crates/cockpit-tui/src/tui/settings/mcp_page.rs
#	crates/cockpit-tui/tests/tui_db_boundary.rs

+++ crates/cockpit-core/src/daemon/server/image_control_mutations.rs
@@ -1098,7 +1098,7 @@ pub(crate) async fn dispatch_image_control_mutation(
                     journal_target,
                     journal_consumed,
                     journal_intended,
-                    current,
+                    current as i64,
                     journal_response,
                     chrono::Utc::now().timestamp_millis()
                 ],
